requirements.txt
langchain
langchain-community
langchain-ollama
langgraph
chromadb
arize-phoenix
openinference-instrumentation-langchain
opentelemetry-sdk
opentelemetry-exporter-otlp
ragas
pandas
python-dotenv
tiktoken
